COMP: Eliminate MSVC71 warnings.

+++ VTK/Utilities/vtkexodus2/excre.c
@@ -48,7 +48,7 @@
 *
 * revision history - 
 *
-*  $Id: excre.c,v 1.3 2006-11-29 18:09:13 dcthomp Exp $
+*  $Id: excre.c,v 1.4 2006-11-29 21:32:36 dcthomp Exp $
 *
 *****************************************************************************/
 #include "exodusII.h"
@@ -177,7 +177,7 @@ int ex_create (const char *path,
  */
 
 /* store Exodus API version # as an attribute */
-  vers = EX_API_VERS;
+  vers = (float)EX_API_VERS;
   if (ncattput (exoid, NC_GLOBAL, ATT_API_VERSION, NC_FLOAT, 1, &vers) == -1)
   {
     exerrval = ncerr;
@@ -189,7 +189,7 @@ int ex_create (const char *path,
   }
 
 /* store Exodus file version # as an attribute */
-  vers = EX_VERS;
+  vers = (float)EX_VERS;
   if (ncattput (exoid, NC_GLOBAL, ATT_VERSION, NC_FLOAT, 1, &vers) == -1)
   {
     exerrval = ncerr;
